Refactor lexer into module & upgrade lexer & parser dependencies (#493)
