+++ deny.toml
@@ -125,7 +125,7 @@ allow = [
 deny = [
     # Use `once_cell` instead
     #   `OnceCell`s API is under consideration for inclusion in `std`: https://github.com/rust-lang/rust/issues/74465
-    { name = "lazy_static", wrappers = ["Inflector", "criterion", "insta", "console"] },
+    { name = "lazy_static", wrappers = ["Inflector", "criterion", "insta", "console", "logos-codegen"] },
     # Advisory: https://rustsec.org/advisories/RUSTSEC-2020-0071
     # `time` < 0.2.23 has a potential (though unlikely) potential segfault
     { name = "time", version = "<0.2.23", wrappers = ["chrono"] },
