+++ partiql-parser/Cargo.toml
@@ -22,7 +22,7 @@ bench = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [build-dependencies]
-lalrpop = "0.20"
+lalrpop = "0.21"
 
 [dependencies]
 partiql-ast = { path = "../partiql-ast", version = "0.10.*" }
@@ -37,8 +37,8 @@ rust_decimal = { version = "1.25.0", default-features = false, features = ["std"
 
 bitflags = "2"
 
-lalrpop-util = "0.20"
-logos = "0.12"
+lalrpop-util = "0.21"
+logos = "0.14"
 
 itertools = "~0.10.3"
